requirements.txt
streamlit==1.40.2
llama-index-core==0.12.16
llama-index-llms-google-genai==0.1.5
llama-index-embeddings-google-genai==0.1.3
llama-index-vector-stores-chroma==0.4.1
chromadb==0.5.23
requests==2.32.3
python-dotenv==1.0.1
